+++ src/redir.h
@@ -49,7 +49,7 @@ struct server
 {
     int fd;
     ssize_t buf_len;
-    int buf_idx;
+    ssize_t buf_idx;
     char *buf; // server send from, remote recv into
     struct sockaddr_in destaddr;
     struct enc_ctx *e_ctx;
@@ -71,7 +71,7 @@ struct remote
 {
     int fd;
     ssize_t buf_len;
-    int buf_idx;
+    ssize_t buf_idx;
     char *buf; // remote send from, server recv into
     struct remote_ctx *recv_ctx;
     struct remote_ctx *send_ctx;
