+++ src/tunnel.h
@@ -53,7 +53,7 @@ struct server
 {
     int fd;
     ssize_t buf_len;
-    int buf_idx;
+    ssize_t buf_idx;
     char *buf; // server send from, remote recv into
     struct enc_ctx *e_ctx;
     struct enc_ctx *d_ctx;
@@ -75,7 +75,7 @@ struct remote
 {
     int fd;
     ssize_t buf_len;
-    int buf_idx;
+    ssize_t buf_idx;
     char *buf; // remote send from, server recv into
     struct remote_ctx *recv_ctx;
     struct remote_ctx *send_ctx;
